refactor: move product-service to a separate folder
