+++ product-service/.npmignore
