+++ product-service/bin/rsschool-aws-be-app.ts
