+++ product-service/lambda_func/add_product.ts
