+++ product-service/lambda_func/products_by_id.ts
