+++ product-service/lambda_func/products_list.ts
