+++ product-service/package.json
@@ -15,7 +15,7 @@
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
-    "@types/node": "20.12.7",
+    "@types/node": "^20.12.7",
     "@types/uuid": "^10.0.0",
     "aws-cdk": "2.145.0",
     "jest": "^29.7.0",
