+++ product-service/swagger.yaml
